Clean up progress providers structure
