+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/core/CSSAnimation.h
@@ -3,7 +3,7 @@
 #include <reanimated/CSS/config/CSSAnimationConfig.h>
 #include <reanimated/CSS/easing/EasingFunctions.h>
 #include <reanimated/CSS/interpolation/styles/AnimationStyleInterpolator.h>
-#include <reanimated/CSS/progress/AnimationProgressProvider.h>
+#include <reanimated/CSS/progress/animation/AnimationProgressProvider.h>
 
 #include <memory>
 #include <string>
