+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/progress/animation/AnimationProgressProvider.h
@@ -3,10 +3,10 @@
 #include <reanimated/CSS/config/CSSAnimationConfig.h>
 #include <reanimated/CSS/config/CSSKeyframesConfig.h>
 #include <reanimated/CSS/easing/EasingFunctions.h>
-#include <reanimated/CSS/progress/KeyframeProgressProvider.h>
-#include <reanimated/CSS/progress/RawProgressProvider.h>
+#include <reanimated/CSS/progress/common/TimeProgressProviderBase.h>
 
 #include <memory>
+#include <utility>
 
 namespace reanimated::css {
 
@@ -17,8 +17,7 @@ enum class AnimationProgressState {
   Finished
 };
 
-class AnimationProgressProvider final : public KeyframeProgressProvider,
-                                        public RawProgressProvider {
+class AnimationProgressProvider final : public TimeProgressProviderBase {
  public:
   AnimationProgressProvider(
       double timestamp,
