+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/progress/common/KeyframeProgressProvider.h
@@ -7,7 +7,6 @@ namespace reanimated::css {
 class KeyframeProgressProvider {
  public:
   virtual double getGlobalProgress() const = 0;
-
   virtual double getKeyframeProgress(double fromOffset, double toOffset)
       const = 0;
 };
