+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/progress/common/TimeProgressProviderBase.cpp
@@ -1,29 +1,29 @@
-#include <reanimated/CSS/progress/RawProgressProvider.h>
+#include <reanimated/CSS/progress/common/TimeProgressProviderBase.h>
 
 namespace reanimated::css {
 
-RawProgressProvider::RawProgressProvider(
+TimeProgressProviderBase::TimeProgressProviderBase(
     const double timestamp,
     const double duration,
     const double delay)
     : duration_(duration), delay_(delay), creationTimestamp_(timestamp) {}
 
-void RawProgressProvider::setDuration(double duration) {
+void TimeProgressProviderBase::setDuration(double duration) {
   resetProgress();
   duration_ = duration;
 }
 
-void RawProgressProvider::setDelay(double delay) {
+void TimeProgressProviderBase::setDelay(double delay) {
   resetProgress();
   delay_ = delay;
 }
 
-void RawProgressProvider::resetProgress() {
+void TimeProgressProviderBase::resetProgress() {
   rawProgress_.reset();
   previousRawProgress_.reset();
 }
 
-void RawProgressProvider::update(const double timestamp) {
+void TimeProgressProviderBase::update(const double timestamp) {
   previousRawProgress_ = rawProgress_;
 
   if (timestamp - creationTimestamp_ < delay_) {
